Rename preferences.py to global_objects.py
Since this file has global PreferenceParser and FontValidator object, this name is more accurate
