+++ modules/global_objects.py
